requirements.txt
fastapi==0.95.2
uvicorn==0.21.1
faiss-cpu==1.7.3
pandas==1.5.3
numpy==1.24.4
sentence-transformers==3.2.1
transformers==4.46.3
torch==2.4.1
